fix: satisfy backend-size and pkg-file-count checks for CLI inputs inventory
