+++ pkg/cli/cliinputs/walk_production_test.go
@@ -1,16 +1,21 @@
 package cliinputs_test
 
 import (
+	"bytes"
 	"os"
+	"reflect"
+	"strings"
 	"testing"
 
 	"github.com/portpowered/infinite-you/pkg/cli"
 	"github.com/portpowered/infinite-you/pkg/cli/cliinputs"
 	"github.com/portpowered/infinite-you/pkg/cli/commandidentity"
 	"github.com/portpowered/infinite-you/pkg/testutil"
+	"github.com/spf13/cobra"
 )
 
 const cliCommandsBaselineFixture = "contracts/testdata/baseline/cli-commands.json"
+const cliCommandInputsBaselineFixture = "contracts/testdata/baseline/cli-command-inputs.json"
 
 func TestWalk_ProductionRootJoinsCommittedCommandIdentityBaseline(t *testing.T) {
 	root := cli.NewRootCommand()
@@ -202,4 +207,269 @@ func flagLongPresent(flags []cliinputs.FlagRecord, longName string) bool {
 		}
 	}
 	return false
+}
+
+func TestWalk_ProductionInventoryMatchesCommittedBaseline(t *testing.T) {
+	root := cli.NewRootCommand()
+
+	inventory, err := cliinputs.Walk(root)
+	if err != nil {
+		t.Fatalf("Walk(production root) error = %v", err)
+	}
+
+	got, err := cliinputs.MarshalInventory(inventory)
+	if err != nil {
+		t.Fatalf("MarshalInventory() error = %v", err)
+	}
+
+	fixturePath := testutil.MustRepoPath(t, cliCommandInputsBaselineFixture)
+	want, err := os.ReadFile(fixturePath)
+	if err != nil {
+		t.Fatalf("read baseline fixture %s: %v", fixturePath, err)
+	}
+
+	if bytes.Equal(got, want) {
+		return
+	}
+
+	t.Fatalf(
+		"CLI command inputs baseline drift detected; update %s when intentional\nwant %d bytes, got %d bytes",
+		cliCommandInputsBaselineFixture,
+		len(want),
+		len(got),
+	)
+}
+
+func TestWriteProductionInputsInventoryBaseline(t *testing.T) {
+	if os.Getenv("UPDATE_CLI_BASELINES") != "1" {
+		t.Skip("set UPDATE_CLI_BASELINES=1 to rewrite fixtures")
+	}
+
+	root := cli.NewRootCommand()
+	inventory, err := cliinputs.Walk(root)
+	if err != nil {
+		t.Fatalf("Walk(production root) error = %v", err)
+	}
+	got, err := cliinputs.MarshalInventory(inventory)
+	if err != nil {
+		t.Fatalf("MarshalInventory() error = %v", err)
+	}
+
+	fixturePath := testutil.MustRepoPath(t, cliCommandInputsBaselineFixture)
+	if err := os.WriteFile(fixturePath, got, 0o644); err != nil {
+		t.Fatalf("write baseline fixture %s: %v", fixturePath, err)
+	}
+}
+
+func TestWalk_ArgumentsSortedByCommandPathPositionName(t *testing.T) {
+	root := newDeterminismArgumentTree()
+
+	inv, err := cliinputs.Walk(root)
+	if err != nil {
+		t.Fatalf("Walk() error = %v", err)
+	}
+
+	for i := 1; i < len(inv.Arguments); i++ {
+		prev := inv.Arguments[i-1]
+		curr := inv.Arguments[i]
+		if !argumentRecordsSorted(prev, curr) {
+			t.Fatalf(
+				"arguments not sorted at index %d: %#v should precede %#v",
+				i,
+				prev,
+				curr,
+			)
+		}
+	}
+}
+
+func TestWalk_FlagsSortedByCommandPathLong(t *testing.T) {
+	root := newSyntheticFlagTree()
+
+	inv, err := cliinputs.Walk(root)
+	if err != nil {
+		t.Fatalf("Walk() error = %v", err)
+	}
+
+	for i := 1; i < len(inv.Flags); i++ {
+		prev := inv.Flags[i-1]
+		curr := inv.Flags[i]
+		if !flagRecordsSorted(prev, curr) {
+			t.Fatalf(
+				"flags not sorted at index %d: %#v should precede %#v",
+				i,
+				prev,
+				curr,
+			)
+		}
+	}
+}
+
+func TestWalk_RelationshipsSortedByCommandPathKindParticipants(t *testing.T) {
+	root := newSyntheticRelationshipTree()
+
+	inv, err := cliinputs.Walk(root)
+	if err != nil {
+		t.Fatalf("Walk() error = %v", err)
+	}
+
+	for i := 1; i < len(inv.Relationships); i++ {
+		prev := inv.Relationships[i-1]
+		curr := inv.Relationships[i]
+		if !relationshipRecordsSorted(prev, curr) {
+			t.Fatalf(
+				"relationships not sorted at index %d: %#v should precede %#v",
+				i,
+				prev,
+				curr,
+			)
+		}
+	}
+}
+
+func TestWalk_DoesNotMutateCommandTree(t *testing.T) {
+	oldSorting := cobra.EnableCommandSorting
+	cobra.EnableCommandSorting = false
+	t.Cleanup(func() { cobra.EnableCommandSorting = oldSorting })
+
+	root := newUnsortedChildOrderTree()
+	beforeOrder := childRegistrationOrder(root)
+
+	_, err := cliinputs.Walk(root)
+	if err != nil {
+		t.Fatalf("Walk() error = %v", err)
+	}
+
+	afterOrder := childRegistrationOrder(root)
+	if !reflect.DeepEqual(beforeOrder, afterOrder) {
+		t.Fatalf("walker mutated child registration order:\nbefore=%#v\nafter=%#v", beforeOrder, afterOrder)
+	}
+}
+
+func TestWalk_ProducesIdenticalJSONOnRepeat(t *testing.T) {
+	root := newDeterminismInventoryTree()
+
+	first, err := cliinputs.Walk(root)
+	if err != nil {
+		t.Fatalf("first Walk() error = %v", err)
+	}
+	firstJSON, err := cliinputs.MarshalInventory(first)
+	if err != nil {
+		t.Fatalf("first MarshalInventory() error = %v", err)
+	}
+
+	second, err := cliinputs.Walk(root)
+	if err != nil {
+		t.Fatalf("second Walk() error = %v", err)
+	}
+	secondJSON, err := cliinputs.MarshalInventory(second)
+	if err != nil {
+		t.Fatalf("second MarshalInventory() error = %v", err)
+	}
+
+	if !bytes.Equal(firstJSON, secondJSON) {
+		t.Fatalf("repeated walks produced different JSON:\nfirst=%s\nsecond=%s", firstJSON, secondJSON)
+	}
+}
+
+func argumentRecordsSorted(left, right cliinputs.ArgumentRecord) bool {
+	if left.CommandPath != right.CommandPath {
+		return left.CommandPath < right.CommandPath
+	}
+	if left.Position != right.Position {
+		return left.Position < right.Position
+	}
+	if left.Name != right.Name {
+		return left.Name < right.Name
+	}
+	return left.IDCandidate < right.IDCandidate
+}
+
+func flagRecordsSorted(left, right cliinputs.FlagRecord) bool {
+	if left.CommandPath != right.CommandPath {
+		return left.CommandPath < right.CommandPath
+	}
+	if left.Long != right.Long {
+		return left.Long < right.Long
+	}
+	return left.IDCandidate < right.IDCandidate
+}
+
+func relationshipRecordsSorted(left, right cliinputs.RelationshipRecord) bool {
+	if left.CommandPath != right.CommandPath {
+		return left.CommandPath < right.CommandPath
+	}
+	if left.Kind != right.Kind {
+		return left.Kind < right.Kind
+	}
+	leftParticipants := strings.Join(left.Participants, "\x00")
+	rightParticipants := strings.Join(right.Participants, "\x00")
+	if leftParticipants != rightParticipants {
+		return leftParticipants < rightParticipants
+	}
+	return left.IDCandidate < right.IDCandidate
+}
+
+func newUnsortedChildOrderTree() *cobra.Command {
+	root := &cobra.Command{Use: "order"}
+	root.AddCommand(
+		&cobra.Command{Use: "z"},
+		&cobra.Command{Use: "a"},
+		&cobra.Command{Use: "m"},
+	)
+	return root
+}
+
+func childRegistrationOrder(root *cobra.Command) map[string][]string {
+	order := make(map[string][]string)
+	captureChildRegistrationOrder(root, order)
+	return order
+}
+
+func captureChildRegistrationOrder(cmd *cobra.Command, order map[string][]string) {
+	children := cmd.Commands()
+	names := make([]string, len(children))
+	for i, child := range children {
+		names[i] = child.Name()
+	}
+	order[cmd.CommandPath()] = names
+	for _, child := range children {
+		captureChildRegistrationOrder(child, order)
+	}
+}
+
+func newDeterminismArgumentTree() *cobra.Command {
+	root := &cobra.Command{Use: "det"}
+
+	root.AddCommand(
+		&cobra.Command{
+			Use:  "zebra <topic>",
+			Args: cobra.ExactArgs(1),
+		},
+		&cobra.Command{
+			Use:  "alpha <topic>",
+			Args: cobra.ExactArgs(1),
+		},
+	)
+	return root
+}
+
+func newDeterminismInventoryTree() *cobra.Command {
+	root := &cobra.Command{Use: "det"}
+	root.PersistentFlags().Bool("shared", false, "shared persistent flag")
+
+	root.AddCommand(
+		&cobra.Command{
+			Use:  "zebra <topic>",
+			Args: cobra.ExactArgs(1),
+		},
+		&cobra.Command{
+			Use:  "alpha [note]",
+			Args: cobra.MaximumNArgs(1),
+		},
+	)
+
+	alpha := root.Commands()[1]
+	alpha.Flags().String("local-only", "", "local flag")
+	return root
 }
